+++ app/src/main/java/com/paw/key/data/remote/datasource/image/ImageDataSource.kt
@@ -3,7 +3,6 @@ package com.paw.key.data.remote.datasource.image
 import com.paw.key.data.dto.image.presigned.ImagePresignedRequestDto
 import com.paw.key.data.dto.image.register.ImageRegisterRequestDto
 import com.paw.key.data.service.image.ImageService
-import okhttp3.RequestBody
 import javax.inject.Inject
 
 class ImageDataSource @Inject constructor(
@@ -19,13 +18,5 @@ class ImageDataSource @Inject constructor(
         dto : ImagePresignedRequestDto
     ) = imageService.presignedImage(
         body = dto
-    )
-
-    suspend fun uploadS3(
-        presignedUrl : String,
-        file : RequestBody
-    ) = imageService.uploadToS3(
-        presignedUrl = presignedUrl,
-        file = file
     )
 }
